refactor: 도메인 객체 domain 패키지 변경
