+++ src/main/java/racingcar/domain/numbergenerator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.numbergenerator;
+package racingcar.domain.numbergenerator;
 
 import java.util.Random;
